+++ App/Shared/SectionRoute.swift
@@ -1,6 +1,7 @@
 import Domain
 
 enum SectionRoute: Hashable {
+    case season
     case bed(Bed.ID)
     case planting(Planting.ID)
     case plant(PlantIdentity)